// ui/src/interfaces/FootballInterface.tsx
export interface PredictionCategoryInterface {
    counts: number;
    preds: number;
    preds_true: number;
    preds_true_percentage: number;
}

export interface PredictionSummaryType {
    counts: number;
    average_score: number;
    home_wins: PredictionCategoryInterface;
    draws: PredictionCategoryInterface;
    away_wins: PredictionCategoryInterface;
    gg: PredictionCategoryInterface;
    ng: PredictionCategoryInterface;
    over15: PredictionCategoryInterface;
    under15: PredictionCategoryInterface;
    over25: PredictionCategoryInterface;
    under25: PredictionCategoryInterface;
    over35: PredictionCategoryInterface;
    under35: PredictionCategoryInterface;
}